Updated doc and small refactor
