+++ Client/VRIS/app/src/main/java/com/valtech/amsterdam/recyclist/RecyclistViewBinder.java
@@ -2,7 +2,7 @@ package com.valtech.amsterdam.recyclist;
 
 import android.view.View;
 
-import com.valtech.amsterdam.vris.ui.OnClickListener;
+import com.valtech.amsterdam.vris.model.OnClickListener;
 
 /**
  * Created by jasper.van.zijp on 26-5-2017.
