+++ Client/VRIS/app/src/main/java/com/valtech/amsterdam/recyclist/modifiers/IHasId.java
@@ -4,6 +4,13 @@ package com.valtech.amsterdam.recyclist.modifiers;
  * Created by marvin.brouwer on 27-7-2017.
  */
 
+/**
+ * Interface indicating this item has an Id getter
+ */
 public interface IHasId {
+    /**
+     * Get the id of this item
+     * @return integer Id
+     */
     int getId();
 }
