+++ Client/VRIS/app/src/main/java/com/valtech/amsterdam/vris/model/Person.java
@@ -4,31 +4,49 @@ package com.valtech.amsterdam.vris.model;
  * Created by jasper.van.zijp on 14-7-2017.
  */
 
-public class Person {
-    private int mId;
-    private String mName;
+import com.google.gson.annotations.SerializedName;
+import com.valtech.amsterdam.recyclist.annotation.ApiInfo;
+import com.valtech.amsterdam.recyclist.modifiers.IHasId;
 
+/**
+ * A person returned by the API
+ */
+@ApiInfo(methodName = "person")
+public final class Person implements IHasId {
+    @SerializedName("Id") private int mId;
+    @SerializedName("Name") private String mName;
+
+    /**
+     * Initiate a new person
+     * @param id
+     * @param name
+     */
     public Person(int id, String name) {
         mId = id;
         mName = name;
     }
 
+    /**
+     * Get the Id of this person
+     * @return
+     */
     public int getId() {
         return mId;
     }
 
-    public void setId(int id) {
-        mId = id;
-    }
-
+    /**
+     * Get the name of this person
+     * @return
+     */
     public String getName() {
         return mName;
     }
+
+    /**
+     * Get the email of this person
+     * @return todo auto generated, needs to be implemented in constructor
+     */
     public String getEmail() {
         return mName + "@email.later.implemented";
-    }
-
-    public void setName(String name) {
-        mName = name;
     }
 }
