+++ Client/VRIS/app/src/main/java/com/valtech/amsterdam/vris/model/Room.java
@@ -4,28 +4,40 @@ package com.valtech.amsterdam.vris.model;
  * Created by jasper.van.zijp on 14-7-2017.
  */
 
-public class Room {
-    private int mId;
-    private String mName;
+import com.google.gson.annotations.SerializedName;
+import com.valtech.amsterdam.recyclist.annotation.ApiInfo;
+import com.valtech.amsterdam.recyclist.modifiers.IHasId;
 
+/**
+ * A room registered in the Office namespace
+ * Todo: unused
+ */
+@ApiInfo(methodName = "room")
+public class Room implements IHasId {
+    @SerializedName("Id") private int mId;
+    @SerializedName("Name")private String mName;
+
+    /**
+     * Initiate a new room
+     * @param id
+     * @param name
+     */
     public Room(int id, String name) {
         mId = id;
         mName = name;
     }
 
+    /**
+     * Get the id of the room
+     * @return
+     */
     public int getId() {
         return mId;
     }
 
-    public void setId(int id) {
-        mId = id;
-    }
-
-    public String getName() {
-        return mName;
-    }
-
-    public void setName(String name) {
-        mName = name;
-    }
+    /**
+     * Get the rooms name
+     * @return
+     */
+    public String getName() { return mName; }
 }
