+++ Client/VRIS/app/src/main/java/com/valtech/amsterdam/vris/model/TimeSlot.java
@@ -13,39 +13,55 @@ import org.joda.time.LocalDateTime;
  * Created by jasper.van.zijp on 18-7-2017.
  */
 
+/**
+ * An empty @ITimeSlot between two @Reservations
+ */
 public class TimeSlot implements ITimeSlot {
     @SerializedName("Id") private int mId;
     @SerializedName("Start") private LocalDateTime mStart;
     @SerializedName("End") private LocalDateTime mEnd;
     private boolean isSelected;
 
+    /**
+     * Initiate a new timeslot
+     * @param id
+     * @param start
+     * @param end
+     */
     public TimeSlot(int id, LocalDateTime start, LocalDateTime end) {
         mId = id;
         mStart = start;
         mEnd = end;
     }
 
+    /**
+     * Get the id of the current TimeSlot
+     * @return
+     */
     public int getId() {
         return mId;
     }
-    public void setId(int id) {
-        mId = id;
-    }
 
+    /**
+     * Get the date the TimeSlot starts in local time
+     * @return
+     */
     public LocalDateTime getStartDate() {
         return mStart;
     }
-    public void setStartDate(LocalDateTime start) {
-        mStart = start;
-    }
 
+    /**
+     * Get the date the TimeSlot end in local time
+     * @return
+     */
     public LocalDateTime getEndDate() {
         return mEnd;
     }
-    public void setEndDate(LocalDateTime end) {
-        mEnd = end;
-    }
 
+    /**
+     * Get the length of the TimeSlot in minutes
+     * @return
+     */
     public int getDurationInMinutes(){
         long diffInMillis =
             mEnd.toDateTime(DateTimeZone.UTC).getMillis() -
@@ -54,13 +70,25 @@ public class TimeSlot implements ITimeSlot {
         return diffInSeconds / 60;
     }
 
+    /**
+     * Get whether or not this TimeSlot is the selected TimeSlot in the menu
+     * @return
+     */
     public boolean getSelected(){
         return isSelected;
     }
+    /**
+     * Set this TimeSlot to be selected
+     * @param isSelected
+     */
     public void setSelected(boolean isSelected){
         this.isSelected = isSelected;
     }
 
+    /**
+     * Serialize the content
+     * @return
+     */
     public ContentValues toContentValues() {
         ContentValues cv = new ContentValues();
         cv.put("Id", getId());
@@ -69,6 +97,10 @@ public class TimeSlot implements ITimeSlot {
         return cv;
     }
 
+    /**
+     * Deserialize the content
+     * @return
+     */
     public static TimeSlot fromContentValues(ContentValues cv) {
         int id = cv.getAsInteger("Id");
         LocalDateTime start = DateTime.parse(cv.getAsString("Start")).toLocalDateTime();
