+++ Client/VRIS/app/src/main/java/com/valtech/amsterdam/vris/ui/BaseFragment.java
@@ -1,18 +1,9 @@
 package com.valtech.amsterdam.vris.ui;
 
-import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 
 /**
  * Created by marvin.brouwer on 20-7-2017.
  */
 
-public abstract class BaseFragment extends Fragment {
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        Log.d("Fragment","onCreate");
-        super.onCreate(savedInstanceState);
-    }
-}
+public abstract class BaseFragment extends Fragment { }
