--- Client/VRIS/app/src/main/java/com/valtech/amsterdam/vris/ui/OnClickListener.java
@@ -1,11 +1,0 @@
-package com.valtech.amsterdam.vris.ui;
-
-import com.valtech.amsterdam.vris.model.ITimeSlot;
-
-/**
- * Created by jasper.van.zijp on 14-7-2017.
- */
-
-public interface OnClickListener {
-    void onClick(ITimeSlot item, int position);
-}
